refactor: move RepoSource into environment
